Rename projectViewModelBuilder to projectTransformer.
